Refactored buildDidDocument
