+++ common-libs/common/test/factory/didFactory.ts
@@ -42,7 +42,7 @@ export const generateTestDIDs = async (): Promise<{
   keysService = new KeysService(joloEncryptedSeed, password)
 
   didDocumentService = DidDocumentService.createDidDocumentService(keysService)
-  const joloDidDocument = await didDocumentService.buildDidDocument(didResolverMock)
+  const joloDidDocument = await didDocumentService.getDidDocument(didResolverMock)
   const joloDid = joloDidDocument.id
 
   const joloPublicKey = KeysService.getPublicKey(joloSeedHex, 'jolo').toString('hex')
@@ -57,7 +57,7 @@ export const generateTestDIDs = async (): Promise<{
   keysService = new KeysService(elemEncryptedSeed, password)
 
   didDocumentService = DidDocumentService.createDidDocumentService(keysService)
-  const elemDidDocument = await didDocumentService.buildDidDocument(didResolverMock)
+  const elemDidDocument = await didDocumentService.getDidDocument(didResolverMock)
   const elemDid = await didDocumentService.getMyDid()
 
   const elemPublicKey = KeysService.getPublicKey(elemSeedHex, 'elem').toString('hex')
@@ -118,7 +118,7 @@ export const generateTestDIDs = async (): Promise<{
   const elemRSAPublicKeyRSA = keysService.getExternalPublicKey('rsa').toString()
 
   didDocumentService = DidDocumentService.createDidDocumentService(keysService)
-  const elemRSADidDocument = await didDocumentService.buildDidDocument(didResolverMock)
+  const elemRSADidDocument = await didDocumentService.getDidDocument(didResolverMock)
   const elemRSADid = await didDocumentService.getMyDid()
 
   const elemRSAPublicKey = KeysService.getPublicKey(elemRSASeedHex, 'elem').toString('hex')
@@ -144,7 +144,7 @@ export const generateTestDIDs = async (): Promise<{
   const elemBBSPublicKeyBBS = keysService.getExternalPublicKey('bbs').toString()
 
   didDocumentService = DidDocumentService.createDidDocumentService(keysService)
-  const elemBBSDidDocument = await didDocumentService.buildDidDocument(didResolverMock)
+  const elemBBSDidDocument = await didDocumentService.getDidDocument(didResolverMock)
   const elemBBSDid = await didDocumentService.getMyDid()
 
   const elemBBSPublicKey = KeysService.getPublicKey(elemBBSSeedHex, 'elem').toString('hex')
